+++ functions/api/image/nfts/collection/collectionImage.test.ts
@@ -23,3 +23,12 @@ test.each(invalidCollectionImageUrls)('invalidAssetImageUrl', async (url) => {
   const response = await fetch(new Request(url))
   expect(response.status).toBeOneOf([404, 500])
 })
+
+const blockedCollectionImageUrls = [
+  'http://127.0.0.1:3000/api/image/nfts/collection/0xd4d871419714b778ebec2e22c7c53572b573706e',
+]
+
+test.each(blockedCollectionImageUrls)('blockedCollectionImageUrl', async (url) => {
+  const response = await fetch(new Request(url))
+  expect(response.status).toBeOneOf([404, 500])
+})
